pyproject.toml
[tool.poetry]
name = "Wp-Quiz-Backend"
version = "0.1.0"
description = "Wp Quiz Backend"
authors = ["Leo Kim <me@leok.kr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
slowapi = "^0.1.5"
pydantic = "^1.8.2"
python-dotenv = "^0.19.1"
aiohttp = "^3.7.4"
gunicorn = "^20.1.0"
motor = "^3.4.0"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
